fix: CourseCard 공통으로 활용
