+++ app/[locale]/academics/undergraduate/courses/CoursePageContent.tsx
@@ -2,7 +2,7 @@
 
 import { useState } from 'react';
 
-import { Course, SortOption, ViewOption } from '@/types/academics';
+import { Classification, Course, SortOption, ViewOption } from '@/types/academics';
 
 import useResponsive from '@/utils/hooks/useResponsive';
 
@@ -42,10 +42,38 @@ export default function CoursePageContent({ courses }: CoursePageContentProps) {
         changeOptions={changeOptions}
       />
       {selectedOption.view === '카드형' ? (
-        <CourseCards courses={courses} selectedOption={selectedOption.sort} />
+        <CourseCards
+          courses={sortCourses(courses, selectedOption.sort)}
+          selectedOption={selectedOption.sort}
+        />
       ) : (
         <CourseList courses={courses} selectedOption={selectedOption.sort} />
       )}
     </>
   );
 }
+
+const getSortGroupIndexByClassification = (classification: Classification) => {
+  if (classification === '전공필수') return 0;
+  else if (classification === '전공선택') return 1;
+  else return 2;
+};
+
+const sortCourses = (courses: Course[], sortOption: SortOption) => {
+  const sortedCourses: Course[][] = [];
+
+  if (sortOption === '학년') {
+    sortedCourses.push([], [], [], []);
+    courses.forEach((course) => sortedCourses[parseInt(course.grade) - 1].push(course));
+  } else if (sortOption === '교과목 구분') {
+    sortedCourses.push([], [], []);
+    courses.forEach((course) =>
+      sortedCourses[getSortGroupIndexByClassification(course.classification)].push(course),
+    );
+  } else {
+    sortedCourses.push([], [], [], []);
+    courses.forEach((course) => sortedCourses[course.credit - 1].push(course));
+  }
+
+  return sortedCourses;
+};
